requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
requests==2.32.3
python-dotenv==1.0.1
psycopg2-binary==2.9.9
Jinja2==3.1.4
alembic==1.13.1
langchain==0.3.7
langchain-openai==0.2.6
pydantic>=2.0.0
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
